Module bam::bgzip

A module that allows to read and write bgzip files directly, as well as modify bgzip blocks.

Re-exports

pub use read::SeekReader;

pub use read::ConsecutiveReader;

pub use read::ReadBgzip;

pub use write::Writer;

pub use write::WriterBuilder;

Modules

read

Bgzip files (BGZF) readers.

write

Bgzip files (BGZF) writer.

Structs

Block

A bgzip block, that can contain compressed, uncompressed data, or both.

Enums

BlockError

Error produced while reading or decompressing a bgzip block.

BlockState

Enum that describes the block state.

Constants

MAX_BLOCK_SIZE

Biggest possible size of the compressed and uncompressed block (= 65536).

MAX_COMPRESSED_SIZE

Biggest possible length of the compressed data (excluding header + footer). Equal to MAX_BLOCK_SIZE - 26 = 65510.