Module bam::bgzip::read

Bgzip files (BGZF) readers.

The module contains two readers: Consecutive Reader and Seek Reader. Both readers have abilities to decompress blocks using additional threads.

Use ReadBgzip trait if you wish to read blocks directly (not via io::Read).

Structs

ConsecutiveReader

Reads bgzip file in a consecutive mode. Therefore, the stream does not have to implement io::Seek.

SeekReader

A bgzip reader, that allows to jump between blocks.

Traits

ReadBgzip

A trait that allows to read blocks directly.