Struct bam::bgzip::read::SeekReader

pub struct SeekReader<R: Read + Seek> { /* fields omitted */ }

A bgzip reader, that allows to jump between blocks.

You can open the reader using from_path or from_stream. Additional threads are used to decompress blocks, while the main thread reads the blocks from a file/stream. If additional_threads is 0, the main thread will decompress blocks itself.

When this reader is opened, it does read not anything, until you specify reading regions. You can do that by using Seek trait, but it is recommended to use methods set_chunks, make_consecutive and from_offset. Note, that in multi-threaded SeekReader it is important to correctly set future reading queue (use set_chunks instead of running from_offset multiple times).

You can read the contents using io::Read, or read blocks using ReadBgzip trait.

Implementations

impl SeekReader<File>

pub fn from_path<P: AsRef<Path>>(
path: P,
additional_threads: u16
) -> Result<Self>

Opens a reader from a file.

impl<R: Read + Seek> SeekReader<R>

pub fn from_stream(stream: R, additional_threads: u16) -> Result<Self>

Opens a reader from a stream.

pub fn set_chunks<I: IntoIterator<Item = Chunk>>(&mut self, chunks: I)

Sets the current chunks. Each chunk specifies a region of the bgzip file. The multi-thread reader reads and decompresses blocks from the chunks in advance (but does not immediately read all blocks).

This function resets the current reading queue (from the previous set_chunks or make_consecutive calls).

chunks must be sorted and should not intersect each other.

pub fn make_consecutive(&mut self)

Sets the reader in a consecutive mode starting with offset 0, and continuing until the end of the stream.

This function resets the current reading queue.

pub fn from_offset(&mut self, offset: VirtualOffset)

Sets the reader in a consecutive mode starting with offset, and continuing until the end of the stream.

This function resets the current reading queue.

pub fn take_stream(self) -> R

Consumes the reader and returns inner stream.

pub fn current_offset(&self) -> VirtualOffset

Returns current virtual offset.

If the reader has not started, returns the start of the first chunk. If the reader finished the current queue, returns the end of the last chunk.