Struct bam::record::sequence::Sequence

pub struct Sequence { /* fields omitted */ }

Wrapper around raw sequence, stored as an [u8; (len + 1) / 2]. Each four bits encode a nucleotide in the following order: =ACMGRSVTWYHKDBN.

Implementations

impl Sequence

pub fn new() -> Self

Creates an empty sequence.

pub fn from_text<I: IntoIterator<Item = u8>>(
sequence: I
) -> Result<Self, String>

Creates a new sequence from text representation.

pub fn clear(&mut self)

Clears sequence but does not touch capacity.

pub fn shrink_to_fit(&mut self)

Shrinks inner vector.

pub fn push(&mut self, nt: u8) -> Result<(), String>

Pushes a single nucleotide to the end.

pub fn extend_from_text<I: IntoIterator<Item = u8>>(
&mut self,
nucleotides: I
) -> Result<(), String>

Extends sequence from the text representation.

pub fn fill_from<R: Read>(
&mut self,
stream: &mut R,
new_len: usize
) -> Result<()>

Clears sequence and fills from a raw stream. new_len represents the number of nucleotides, not the number of bytes.

pub fn raw(&self) -> &[u8]

Returns raw data.

pub fn len(&self) -> usize

Returns full length of the sequence, O(1).

pub fn available(&self) -> bool

Returns true, if the sequence is present.

pub fn to_vec(&self) -> Vec<u8>

Returns transformed data, each byte represents a single nucleotde, O(n).

pub fn to_vec_acgtn_only(&self) -> Vec<u8>

Returns transformed data using only nucleotides A, C, G, T and N, all other values are transformed into N, each byte represents a single nucleotde, O(n).

pub fn at(&self, index: usize) -> u8

Returns a nucleotide at the position index, represented by a single byte, O(1).

pub fn at_acgtn_only(&self, index: usize) -> u8

Returns a nucleotide at the position index, represented by a single byte, O(1). If the nucleotide is not A, C, G or T, the function returns N.

pub fn compl_at(&self, index: usize) -> u8

Returns a nucleotide, complement to the nucleotide at the position index, O(1).

pub fn compl_at_acgtn_only(&self, index: usize) -> u8

Returns a nucleotide, complement to the nucleotide at the position index, O(1). If the nucleotide is not A, C, G or T, the function returns N.

pub fn subseq<'a, R: RangeBounds<usize>>(&'a self, range: R) -> SubseqIter<'a>

Returns an iterator over a subsequence.

pub fn subseq_acgtn_only<R: RangeBounds<usize>>(
&self,
range: R
) -> SubseqIterAcgtn<'_>

Returns an iterator over a subsequence using only nucleotides A, C, G, T and N.

pub fn rev_compl<R: RangeBounds<usize>>(&self, range: R) -> RevComplIter<'_>

Returns an iterator over a reverse complement of a subsequence.

pub fn rev_compl_acgtn_only<R: RangeBounds<usize>>(
&self,
range: R
) -> RevComplIterAcgtn<'_>

Returns an iterator over a reverse complement of a subsequence using only nucleotides A, C, G, T and N.

pub fn write_readable<W: Write>(&self, f: &mut W) -> Result<()>

Writes in human readable format. Writes * if empty.