Module bam::record::tags

Record tags and operations on them.

Structs

FloatArrayView

Wrapper around raw float array stored in a tag.

FloatArrayViewIter

Double-ended iterator over FloatArrayView values.

IntArrayView

Wrapper around raw integer array stored in a tag.

IntArrayViewIter

Double-ended iterator over IntArrayView values converted into i64.

TagIter

Iterator over tags.

TagViewer

Wrapper around raw tags.

Enums

IntegerType

Enum that represents tag type for the cases when a tag contains integer.

StringType

Enum that represents tag type for String and Hex types (Z and H).

TagValue

Enum with all possible tag values.

Traits

PushNum

Type Definitions

TagName

Alias for a tag name.